Separated config commands in sub group for platform and deployment
