+++ Babylon/groups/config/groups/__init__.py
@@ -1,3 +1,7 @@
+from .platform import platform
+from .deployment import deployment
 
 list_groups = [
+    platform,
+    deployment,
 ]
