+++ Babylon/groups/config/groups/deployment/groups/__init__.py
@@ -1,0 +1,3 @@
+
+list_groups = [
+]
