+++ Babylon/groups/config/groups/platform/__init__.py
@@ -1,0 +1,20 @@
+from click import group
+from click import pass_context
+from click.core import Context
+
+from .commands import list_commands
+from .groups import list_groups
+
+
+@group()
+@pass_context
+def platform(ctx: Context):
+    """Group initialized from a template"""
+    pass
+
+
+for _command in list_commands:
+    platform.add_command(_command)
+
+for _group in list_groups:
+    platform.add_command(_group)
