+++ Babylon/groups/config/groups/platform/commands/__init__.py
@@ -1,0 +1,9 @@
+from .create import create
+from .edit import edit
+from .select import select
+
+list_commands = [
+    create,
+    edit,
+    select,
+]
